Move filter expression helpers to DuckLakeUtil
